requirements.txt
# Python dependencies for the Telegram bridge.
# Browser automation uses the external `agent-browser` CLI:
# https://agent-browser.dev
# It is not installed from this requirements file.
requests>=2.31.0
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
